refactor(logs): lift OTLP utilities to @posthog/core
Move OTLP serialization helpers from packages/browser/src/logs-utils.ts into
packages/core/src/logs/logs-utils.ts so the upcoming React Native logs
feature consumes the same builders. Browser now imports them from
@posthog/core; the local browser file and its test are deleted.

User-visible behavior change for browser:
NaN and plus or minus Infinity attribute values now serialize as
{stringValue: NaN or Infinity or -Infinity} instead of being silently lost
during JSON encoding. JSON.stringify turns non-finite floats into null,
which the browser SDK was unknowingly emitting on the wire — a silent
data-loss bug. proto3 JSON mapping (which OTLP HTTP rides) requires the
literal strings; this change preserves the value regardless of what
downstream parsers do with null.

Type changes:
- @posthog/types LogSdkContext gains optional screenName and appState
  fields so a single context shape works for both web (currentUrl) and
  mobile (screenName + appState) callers. Existing fields unchanged.
- @posthog/core gains a workspace dependency on @posthog/types so the
  lifted helpers can consume the shared log type vocabulary.
- @posthog/react-native adds @posthog/types as a workspace dep in
  preparation for the upcoming logs module.

Mechanical refactor for browser:
buildOtlpLogsPayload previously hardcoded Config.LIB_NAME inside the
function; the lifted version takes scopeName as a parameter and the
browser caller passes Config.LIB_NAME explicitly. Wire output is unchanged
(same string, same field).
