+++ compliance/node/Dockerfile
@@ -6,6 +6,7 @@ WORKDIR /app
 COPY package.json pnpm-lock.yaml pnpm-workspace.yaml ./
 COPY patches ./patches
 COPY tooling ./tooling
+COPY packages/types ./packages/types
 COPY packages/core ./packages/core
 COPY packages/node ./packages/node
 
@@ -15,7 +16,8 @@ RUN npm install -g pnpm
 # Install dependencies
 RUN pnpm install --frozen-lockfile
 
-# Build the core and node packages
+# Build the types, core, and node packages
+RUN pnpm --filter @posthog/types build
 RUN pnpm --filter @posthog/core build
 RUN pnpm --filter posthog-node build
