+++ packages/core/src/index.ts
@@ -2,6 +2,14 @@ export { getFeatureFlagValue } from './featureFlagUtils'
 export { gzipCompress, isNativeAsyncGzipReadError } from './gzip'
 export * from './utils'
 export * as ErrorTracking from './error-tracking'
+export {
+  buildOtlpLogRecord,
+  buildOtlpLogsPayload,
+  getOtlpSeverityNumber,
+  getOtlpSeverityText,
+  toOtlpAnyValue,
+  toOtlpKeyValueList,
+} from './logs/logs-utils'
 export { uuidv7 } from './vendor/uuidv7'
 export * from './posthog-core'
 export * from './posthog-core-stateless'
